captived: consolidate more of the json parsing and generation
The primary purpose of this commit is changing the resource request
and response body types from std::string to a json type. This allows
all of the json<->string encoding and decoding to be done on spot, and
not repeated across all of the handler functions.

As part of this change, calls to native C libjansson functions are
replaced with calls to the C++ wrappers defined in json.hpp.  This
removes a bunch of C<->C++ interface code and simplifies memory
management.