xbridge/gadget: move usb0 up/down control to xbridge
The xap-usb-gadget service loads the USB ethernet gadget, which
creates the usb0 interface, and then bring the interface up using
`ip`.  bring that interface up. However, there is some sort of race
condition, because sometimes the interface is not actually brought up.
The xbridge service is then unable to start.

To fix this is a safe way, bring the interface up in the xbridge
service.

Since there is a known race condition, also add a small delay after
loading the kernel module.