wpa-credentials: add xaprc_recovery network
The `xaprc_recovery` network will be hardcoded into all router cards
as a failsafe form of last-mile access. If a customer send a bad wifi
config that prevents the router card from reconnecting, they can
create an access point with the xaprc_recovery SSID to allow the
router card to reconnect and download an updated, corrected wifi
config.