wpa-credentials: configure connman to ignore usb0 and enfX interfaces
By default, connman will try to control the usb0, enf0, and enf1
interfaces.  This has normally worked ok, but it can cause problems.
For example, if connman is restarted or stopped, it will change each
of the interface states to "down", which breaks forwarding.

So, prevent these potential problems by instructing connman to
completely ignore interfaces that start with usb or enf.